scripts/requirements-resume.txt
fpdf2>=2.7.9
PyYAML>=6.0.2
